benchmarks/cases/django-invoices-basic/repo/accounts/views.py
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response
from rest_framework.views import APIView


class ProfileView(APIView):
  permission_classes = [IsAuthenticated]

  def get(self, request):
    return Response({
      "id": request.user.id,
      "email": request.user.email
    })
