benchmarks/cases/django-invoices-basic/repo/requirements.txt
django==5.0.2
djangorestframework==3.15.1
